requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.6
xmltodict==0.12.0
